Extract a utility.
