+++ base/src/test/java/io/spine/base/EnvironmentTest.java
@@ -240,74 +240,6 @@ class EnvironmentTest extends UtilityClassTest<Environment> {
         assertThat(environment.type()).isInstanceOf(Local.class);
     }
 
-    @Nested
-    @DisplayName("disallow to register types by class if")
-    class ProhibitRegistering {
-
-        @Test
-        @DisplayName("they have a public ctor")
-        void registerCustomPublicCtor() {
-            assertThrows(IllegalArgumentException.class, () -> environment.register(Travis.class));
-        }
-
-        @Test
-        @DisplayName("they have a private ctor")
-        void registerCustomPrivateCtor() {
-            assertThrows(IllegalArgumentException.class,
-                         () -> environment.register(HiddenEnvironment.class));
-        }
-
-        @Test
-        @DisplayName("they have a protected ctor")
-        void registerCustomPackagePrivateCtor() {
-            assertThrows(IllegalArgumentException.class,
-                         () -> environment.register(ProtectedEnvironment.class));
-        }
-
-        @Test
-        @DisplayName("they do not have a parameterless ctor")
-        void registerCustomCtorWithParameters() {
-            assertThrows(IllegalArgumentException.class,
-                         () -> environment.register(ValueDependantEnvironment.class));
-        }
-
-        private class ValueDependantEnvironment extends EnvironmentType {
-
-            private final boolean enabled;
-
-            ValueDependantEnvironment(boolean enabled) {
-                this.enabled = enabled;
-            }
-
-            @Override
-            protected boolean enabled() {
-                return enabled;
-            }
-        }
-
-        private class HiddenEnvironment extends EnvironmentType {
-
-            private HiddenEnvironment() {
-            }
-
-            @Override
-            protected boolean enabled() {
-                return false;
-            }
-        }
-
-        private class ProtectedEnvironment extends EnvironmentType {
-
-            protected ProtectedEnvironment() {
-            }
-
-            @Override
-            protected boolean enabled() {
-                return false;
-            }
-        }
-    }
-
     static final class Local extends EnvironmentType {
 
         /**
